fix: resolve clippy warnings and code issues across modules
- evo/mod.rs: allow too_many_arguments/type_complexity for KigaliGenome,
  allow borrowed_box for sort_by closures (false positives), fix range loop
- loss/mod.rs: add Default impls for TestLossFunction and TairuaLoss,
  replace map_or with is_some_and, fix range loop in ToneholeTuningLoss
- sim/mod.rs: fix needless_borrow in viscothermal_loss_params call,
  change public API from &Vec<[f64;2]> to &[[f64;2]] (from_geo/with_strategy)
- persistence/mod.rs: replace manual clamp with .clamp(), derive Default for ProjectState
- visualization/mod.rs: remove useless format! calls
- audio/mod.rs: replace clone with deref for Copy type
- waveguide/mod.rs: remove unnecessary f64 cast
- All 55 unit + 14 integration tests pass, cargo clippy clean

+++ src/waveguide/mod.rs
@@ -70,7 +70,7 @@ impl WaveguideCell {
 
     /// Compute the delay in samples for a given sampling rate
     pub fn delay_samples(&self, sample_rate: f64) -> usize {
-        let wavelength = C / (sample_rate as f64 / 1024.0); // Approximate wavelength
+        let wavelength = C / (sample_rate / 1024.0); // Approximate wavelength
         (self.length / wavelength).ceil() as usize
     }
 }
